+++ NEWS.md
@@ -21,6 +21,8 @@
 * Add a new `tar_pattern()` function to emulate dynamic branching outside a pipeline.
 * Add a new `level_separation` argument to `tar_visnetwork()` and `tar_glimpse()` to control the aspect ratio (#226).
 * Track functions from multiple packages with the `imports` argument to `tar_option_set()` (#239).
+* Add color for "built" progress if `outdated` is `FALSE` in `tar_visnetwork()`.
+* Tweak colors in `tar_visnetwork()` to try to account for color blindness.
 
 ## Enhancements